Move the device manager to its own file.
