+++ tests/test_devicemanager.py
@@ -7,6 +7,10 @@ import inputs
 from unittest import mock
 from pathlib import PurePath
 
+from inputs.errors import UnknownEventType
+from inputs.manager import DeviceManager
+
+
 RAW = ""
 
 # Mocking adds an argument, whether we need it or not.
@@ -21,11 +25,11 @@ OTHER_PATH = "/dev/input/by-path/the-machine-that-goes-ping-other"
 class DeviceManagePostrInitTestCase(TestCase):
     """Test the device manager class' post-init method."""
 
-    @mock.patch.object(inputs.DeviceManager, '_find_devices')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_mac')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_win')
-    @mock.patch.object(inputs.DeviceManager, '_find_leds')
-    @mock.patch.object(inputs.DeviceManager, '_update_all_devices')
+    @mock.patch.object(DeviceManager, '_find_devices')
+    @mock.patch.object(DeviceManager, '_find_devices_mac')
+    @mock.patch.object(DeviceManager, '_find_devices_win')
+    @mock.patch.object(DeviceManager, '_find_leds')
+    @mock.patch.object(DeviceManager, '_update_all_devices')
     def test_post_init_linux(
             self,
             mock_update_all_devices,
@@ -34,59 +38,59 @@ class DeviceManagePostrInitTestCase(TestCase):
             mock_find_devices_mac,
             mock_find_devices):
         """On Linux, find_devices is called and the other methods are not."""
-        inputs.NIX = True
-        inputs.WIN = False
-        inputs.MAC = False
+        inputs.manager.NIX = True
+        inputs.manager.WIN = False
+        inputs.manager.MAC = False
         # pylint: disable=unused-variable
-        device_manger = inputs.DeviceManager()
+        device_manger = DeviceManager()
         mock_update_all_devices.assert_called()
         mock_find_devices.assert_called()
         mock_find_devices_mac.assert_not_called()
         mock_find_devices_win.assert_not_called()
         mock_find_leds.assert_called()
 
-    @mock.patch.object(inputs.DeviceManager, '_find_devices')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_mac')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_win')
-    @mock.patch.object(inputs.DeviceManager, '_update_all_devices')
+    @mock.patch.object(DeviceManager, '_find_devices')
+    @mock.patch.object(DeviceManager, '_find_devices_mac')
+    @mock.patch.object(DeviceManager, '_find_devices_win')
+    @mock.patch.object(DeviceManager, '_update_all_devices')
     def test_post_init_mac(self,
                            mock_update_all_devices,
                            mock_find_devices_win,
                            mock_find_devices_mac,
                            mock_find_devices):
         """On Mac, find_devices_mac is called and other methods are not."""
-        inputs.NIX = False
-        inputs.WIN = False
-        inputs.MAC = True
-        inputs.DeviceManager()
+        inputs.manager.NIX = False
+        inputs.manager.WIN = False
+        inputs.manager.MAC = True
+        DeviceManager()
         mock_update_all_devices.assert_called()
         mock_find_devices_mac.assert_called()
         mock_find_devices.assert_not_called()
         mock_find_devices_win.assert_not_called()
 
-    @mock.patch.object(inputs.DeviceManager, '_find_devices')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_mac')
-    @mock.patch.object(inputs.DeviceManager, '_find_devices_win')
-    @mock.patch.object(inputs.DeviceManager, '_update_all_devices')
+    @mock.patch.object(DeviceManager, '_find_devices')
+    @mock.patch.object(DeviceManager, '_find_devices_mac')
+    @mock.patch.object(DeviceManager, '_find_devices_win')
+    @mock.patch.object(DeviceManager, '_update_all_devices')
     def test_post_init_win(self,
                            mock_update_all_devices,
                            mock_find_devices_win,
                            mock_find_devices_mac,
                            mock_find_devices):
         """On Windows, find_devices_win is called and other methods are not."""
-        inputs.WIN = True
-        inputs.MAC = False
-        inputs.NIX = False
-        inputs.DeviceManager()
+        inputs.manager.WIN = True
+        inputs.manager.MAC = False
+        inputs.manager.NIX = False
+        DeviceManager()
         mock_update_all_devices.assert_called()
         mock_find_devices_win.assert_called()
         mock_find_devices.assert_not_called()
         mock_find_devices_mac.assert_not_called()
 
     def tearDown(self):
-        inputs.WIN = False
-        inputs.MAC = False
-        inputs.NIX = True
+        inputs.manager.WIN = False
+        inputs.manager.MAC = False
+        inputs.manager.NIX = True
 
 
 MOCK_DEVICE = 'My Special Mock Input Device'
@@ -100,9 +104,9 @@ class DeviceManagerTestCase(TestCase):
     # There can never be too many tests.
     # pylint: disable=too-many-public-methods
 
-    @mock.patch.object(inputs.DeviceManager, '_post_init')
+    @mock.patch.object(DeviceManager, '_post_init')
     def setUp(self, mock_method):
-        self.device_manger = inputs.DeviceManager()
+        self.device_manger = DeviceManager()
         self.mock_method = mock_method
 
     def test_init(self):
@@ -142,7 +146,7 @@ class DeviceManagerTestCase(TestCase):
         self.assertEqual(self.device_manger.all_devices, [])
 
     @mock.patch('os.path.realpath')
-    @mock.patch('inputs.Keyboard')
+    @mock.patch('inputs.manager.Keyboard')
     def test_parse_device_path_keyboard(
             self,
             mock_keyboard,
@@ -160,7 +164,7 @@ class DeviceManagerTestCase(TestCase):
         self.assertEqual(self.device_manger._raw[0], KEYBOARD_PATH)
 
     @mock.patch('os.path.realpath')
-    @mock.patch('inputs.Keyboard')
+    @mock.patch('inputs.manager.Keyboard')
     def test_parse_device_path_repeated(
             self,
             mock_keyboard,
@@ -179,7 +183,7 @@ class DeviceManagerTestCase(TestCase):
         self.assertEqual(len(self.device_manger.keyboards), 1)
 
     @mock.patch('os.path.realpath')
-    @mock.patch('inputs.Mouse')
+    @mock.patch('inputs.manager.Mouse')
     def test_parse_device_path_mouse(
             self,
             mock_mouse,
@@ -197,7 +201,7 @@ class DeviceManagerTestCase(TestCase):
         self.assertEqual(self.device_manger._raw[0], MOUSE_PATH)
 
     @mock.patch('os.path.realpath')
-    @mock.patch('inputs.GamePad')
+    @mock.patch('inputs.manager.GamePad')
     def test_parse_device_path_gamepad(
             self,
             mock_gamepad,
@@ -215,7 +219,7 @@ class DeviceManagerTestCase(TestCase):
         self.assertEqual(self.device_manger._raw[0], GAMEPAD_PATH)
 
     @mock.patch('os.path.realpath')
-    @mock.patch('inputs.OtherDevice')
+    @mock.patch('inputs.manager.OtherDevice')
     def test_parse_device_path_other(
             self,
             mock_other,
@@ -252,7 +256,7 @@ class DeviceManagerTestCase(TestCase):
 
     def test_get_invalid_event_type(self):
         """get_event_type raises exception for an invalid event type."""
-        with self.assertRaises(inputs.UnknownEventType):
+        with self.assertRaises(UnknownEventType):
             self.device_manger.get_event_type(0x64)
 
     def test_get_event_string(self):
@@ -284,19 +288,19 @@ class DeviceManagerTestCase(TestCase):
 
     def test_get_event_string_on_win(self):
         """get_event_string returns an event string on Windows."""
-        inputs.WIN = True
+        inputs.manager.WIN = True
         self.assertEqual(
             self.device_manger.get_event_string('Key', 0x133),
             "BTN_NORTH")
-        inputs.WIN = False
+        inputs.manager.WIN = False
 
     def test_invalid_event_string(self):
         """get_event_string raises an exception for an unknown event code."""
         with self.assertRaises(inputs.UnknownEventCode):
             self.device_manger.get_event_string('Key', 0x999)
 
-    @mock.patch.object(inputs.DeviceManager, '_find_special')
-    @mock.patch.object(inputs.DeviceManager, '_find_by')
+    @mock.patch.object(DeviceManager, '_find_special')
+    @mock.patch.object(DeviceManager, '_find_by')
     def test_find_devices(self, mock_find_by, mock_find_special):
         """It should find by path, id and specials."""
         self.device_manger._find_devices()
@@ -322,8 +326,8 @@ class DeviceManagerTestCase(TestCase):
         for device in (0, 1, 2, 3, 4):
             self.assertEqual(self.device_manger[device], device)
 
-    @mock.patch.object(inputs.DeviceManager, '_parse_device_path')
-    @mock.patch('inputs.open', mock.mock_open(read_data=MOCK_DEVICE))
+    @mock.patch.object(DeviceManager, '_parse_device_path')
+    @mock.patch('inputs.manager.open', mock.mock_open(read_data=MOCK_DEVICE))
     @mock.patch('glob.glob')
     def test_find_special(self, mock_glob, mock_parse_device_path):
         """Find a special device."""
@@ -351,9 +355,9 @@ class DeviceManagerTestCase(TestCase):
 
             self.assertEqual(device_path, target_path)
 
-    @mock.patch.object(inputs.DeviceManager, '_parse_device_path')
-    @mock.patch.object(inputs.DeviceManager, '_get_char_names')
-    @mock.patch('inputs.open', mock.mock_open(read_data=MOCK_DEVICE))
+    @mock.patch.object(DeviceManager, '_parse_device_path')
+    @mock.patch.object(DeviceManager, '_get_char_names')
+    @mock.patch('inputs.manager.open', mock.mock_open(read_data=MOCK_DEVICE))
     @mock.patch('glob.glob')
     def test_find_special_repeated(self,
                                    mock_glob,
@@ -369,7 +373,7 @@ class DeviceManagerTestCase(TestCase):
         mock_parse_device_path.assert_not_called()
 
     @mock.patch('glob.glob')
-    @mock.patch.object(inputs.DeviceManager, '_parse_device_path')
+    @mock.patch.object(DeviceManager, '_parse_device_path')
     def test_find_by(self,
                      mock_parse_device_path,
                      mock_glob):
@@ -390,14 +394,14 @@ class DeviceManagerPlatformTestCase(TestCase):
     # There can never be too many tests.
     # pylint: disable=too-many-public-methods
 
-    @mock.patch.object(inputs.DeviceManager, '_post_init')
+    @mock.patch.object(DeviceManager, '_post_init')
     def setUp(self, mock_method):
-        self.device_manager = inputs.DeviceManager()
+        self.device_manager = DeviceManager()
         self.mock_method = mock_method
 
-    @mock.patch('inputs.Mouse')
-    @mock.patch('inputs.MightyMouse')
-    @mock.patch('inputs.Keyboard')
+    @mock.patch('inputs.manager.Mouse')
+    @mock.patch('inputs.manager.MightyMouse')
+    @mock.patch('inputs.manager.Keyboard')
     def test_find_devices_mac(self, mock_kb, mock_mighty, mock_mouse):
         """Test the mac version of _find_devices_mac."""
         self.device_manager._find_devices_mac()
@@ -413,7 +417,7 @@ class DeviceManagerPlatformTestCase(TestCase):
         mock_mighty.assert_called_once_with(self.device_manager)
         mock_mouse.assert_called_once_with(self.device_manager)
 
-    @mock.patch('inputs.ctypes.windll',
+    @mock.patch('inputs.manager.ctypes.windll',
                 create=True)
     def test_find_xinput(self, mock_windll):
         """Finds an xinput library if one is available. """
@@ -422,8 +426,8 @@ class DeviceManagerPlatformTestCase(TestCase):
                     str(self.device_manager.xinput._extract_mock_name())
         self.assertTrue(found_one)
 
-    @mock.patch('inputs.XINPUT_DLL_NAMES')
-    @mock.patch('inputs.ctypes.windll',
+    @mock.patch('inputs.manager.XINPUT_DLL_NAMES')
+    @mock.patch('inputs.manager.ctypes.windll',
                 create=True)
     def test_find_xinput_not_available(
             self,
@@ -435,11 +439,11 @@ class DeviceManagerPlatformTestCase(TestCase):
 
         self.assertIsNone(self.device_manager.xinput)
 
-    @mock.patch.object(inputs.DeviceManager, '_find_xinput')
-    @mock.patch.object(inputs.DeviceManager, '_detect_gamepads')
-    @mock.patch.object(inputs.DeviceManager, '_count_devices')
-    @mock.patch('inputs.Mouse')
-    @mock.patch('inputs.Keyboard')
+    @mock.patch.object(DeviceManager, '_find_xinput')
+    @mock.patch.object(DeviceManager, '_detect_gamepads')
+    @mock.patch.object(DeviceManager, '_count_devices')
+    @mock.patch('inputs.manager.Mouse')
+    @mock.patch('inputs.manager.Keyboard')
     def test_find_devices_win(self,
                               mock_keyboard,
                               mock_mouse,
@@ -455,8 +459,8 @@ class DeviceManagerPlatformTestCase(TestCase):
         self.assertTrue(len(self.device_manager.mice) == 1)
         self.assertTrue(len(self.device_manager.keyboards) == 1)
 
-    @mock.patch('inputs.GamePad')
-    @mock.patch('inputs.ctypes.windll', create=True)
+    @mock.patch('inputs.manager.GamePad')
+    @mock.patch('inputs.manager.ctypes.windll', create=True)
     def test_detect_gamepads(self,
                              mock_windll,
                              mock_gamepad):
@@ -468,8 +472,8 @@ class DeviceManagerPlatformTestCase(TestCase):
         self.device_manager._detect_gamepads()
         self.assertEqual(len(self.device_manager.gamepads), 4)
 
-    @mock.patch('inputs.GamePad')
-    @mock.patch('inputs.ctypes.windll', create=True)
+    @mock.patch('inputs.manager.GamePad')
+    @mock.patch('inputs.manager.ctypes.windll', create=True)
     def test_detect_error_gamepads(self,
                                    mock_windll,
                                    mock_gamepad):
@@ -482,7 +486,7 @@ class DeviceManagerPlatformTestCase(TestCase):
             self.device_manager._detect_gamepads()
         self.assertEqual(len(self.device_manager.gamepads), 0)
 
-    @mock.patch('inputs.ctypes.windll', create=True)
+    @mock.patch('inputs.manager.ctypes.windll', create=True)
     def test_count_devices(self, mock_windll):
         """It should count the attached devices."""
         self.device_manager._raw_device_counts = {
